+++ src/main/java/ru/barabanra/yandex/dto/item/weather/Night.java
@@ -8,7 +8,7 @@ public class Night {
 
     private String _source;
 
-    private Long cloudness;
+    private Double cloudness;
 
     private String condition;
